# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Xalendar Official API docs"
  },
  "host": "165.22.72.61",
...
  "paths": {
...
    "/event/{id}": {
      "patch": {
        "tags": ["Event"],
        "summary": "Update specified event",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id of the users",
            "required": true,
            "type": "string"
          },

          {
            "in": "body",
            "name": "body",
            "description": "Updated json",
            "schema": {
              "example": {
                "name": "Практика IT",
                "color": "#696969",
                "aud": "Онлайн",
                "link": "www.discord.com",
                "theme": "Работа с Angular.js",
                "startTime": "2021-05-28T8:30:00.000Z",
                "endTime": "2021-05-28T10:15:00.000Z"
              }
            }
          }
        ],

        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "example": {
                "id": 1,
                "name": "Практика IT",
                "color": "#696969",
                "aud": "Онлайн",
                "link": "www.discord.com",
                "teacher": "Воробьев А.И.",
                "theme": "Работа с Angular.js",
                "start_time": "2021-05-29 14:15:00",
                "end_time": "2021-05-29 15:50:00",
                "module_id": 1
              }
            }
          }
        }
      },

      "delete": {
        "tags": ["Event"],
        "summary": "Delete specified event from the system.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Event id",
            "required": true,
            "type": "string"
          }
        ],

        "responses": {
          "204": {
            "description": "No content",
            "schema": {
              "example": {}
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}